requirements.txt
androguard==3.3.5
networkx==2.2
Pillow==6.2.0
websocket-client==0.54.0
pytest-cov==2.6.1
selenium==3.141.0
termcolor==1.1.0
gym==0.16.0
DateTime==4.3
Appium-Python-Client==0.48
loguru==0.3.2
flasgger==0.9.2
Flask==1.0.2
requests==2.22.0
frida-tools==6.0.1
tensorflow==2.0.0
scikit-learn==0.21.2
numpy==1.18.1
beautifulsoup4==4.8.2
slackclient==1.3.0
tqdm==4.43.0
pandas==1.0.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
langdetect==1.0.8
nltk==3.4.5
xlsxwriter==1.2.8
xlrd==1.2.0
ggplot==0.11.5
google-play-scraper==0.0.2.2
